pyproject.toml
[project]
name = "forex-forcast"
version = "0.1.0"
description = ""
authors = [
    {name = "brink",email = "brinkevich06@gmail.com "}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "yfinance (>=1.2.0,<2.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "numpy (>=2.4.3,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "statsmodels (>=0.14.6,<0.15.0)",
    "arch (>=8.0.0,<9.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "xgboost (>=3.2.0,<4.0.0)",
    "tensorflow (>=2.21.0,<3.0.0)",
    "shap (>=0.51.0,<0.52.0)",
    "fredapi (>=0.5.2,<0.6.0)",
    "scipy (>=1.17.1,<2.0.0)",
    "rich (>=14.3.3,<15.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
